clinical/requirements.txt
absl-py==2.1.0
albucore==0.0.19
albumentations==1.4.20
alembic==1.15.2
anndata==0.11.3
annotated-types==0.7.0
annoy==1.17.3
anyio==4.6.2.post1
array_api_compat==1.11.2
asttokens==3.0.0
attrs==24.2.0
blinker==1.9.0
blis==1.2.1
bokeh==3.6.1
Bottleneck==1.4.2
cachetools==5.5.2
catalogue==2.0.10
catboost==1.2.8
category_encoders==2.7.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
cloudpathlib==0.21.0
cloudpickle==3.1.1
colorama==0.4.6
colorlog==6.9.0
colorlover==0.3.0
comm==0.2.2
confection==0.1.5
contourpy==1.3.2
cufflinks==0.17.3
cycler==0.12.1
cymem==2.0.11
Cython==3.1.0
dash==3.0.4
databricks-sdk==0.53.0
decorator==5.2.1
decoupler==1.8.0
Deprecated==1.2.18
deprecation==2.1.0
docker==7.1.0
eval_type_backport==0.2.2
executing==2.2.0
fastapi==0.115.5
fastapi-cli==0.0.5
fastjsonschema==2.21.1
fbpca==1.0
filelock==3.18.0
fire==0.7.0
Flask==3.0.3
fonttools==4.58.0
fsspec==2025.3.2
funcy==2.0
gensim==4.3.2
geosketch==1.3
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.40.1
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
graphviz==0.20.3
greenlet==3.2.2
h11==0.14.0
h5py==3.12.1
hdf5plugin==5.1.0
htmlmin==0.1.12
httptools==0.6.4
idna==3.10
igraph==0.11.8
ImageHash==4.3.2
imageio==2.37.0
imbalanced-learn==0.13.0
importlib_metadata==8.6.1
importlib_resources==6.5.2
intervaltree==3.1.0
ipython==9.2.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
joblib==1.3.2
json-tricks==3.17.3
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter_core==5.7.2
jupyterlab_widgets==3.0.15
kaleido==0.2.1
kiwisolver==1.4.8
kmodes==0.12.2
langcodes==3.5.0
language_data==1.3.0
lazy_loader==0.4
legacy-api-wrap==1.4.1
leidenalg==0.10.2
lightgbm==4.6.0
llvmlite==0.43.0
Mako==1.3.10
marisa-trie==1.2.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.7.5
matplotlib-inline==0.1.7
mdurl==0.1.2
missingno==0.5.2
mlflow==2.17.2
mlflow-skinny==2.17.2
mlxtend==0.23.4
monai==1.4.0
mpmath==1.3.0
multimethod==1.4
murmurhash==1.0.12
natsort==8.4.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nibabel==5.3.2
nltk==3.9.1
numba==0.60.0
numexpr==2.10.2
numpy==1.26.4
opencv-python-headless==4.10.0.84
opentelemetry-api==1.33.0
opentelemetry-sdk==1.33.0
opentelemetry-semantic-conventions==0.54b0
optuna==4.1.0
orjson==3.10.18
packaging==24.1
pandas==2.1.4
pandas-profiling==3.0.0
parso==0.8.4
patsy==1.0.1
phik==0.12.4
pillow==10.4.0
pip==25.1
platformdirs==4.3.8
plotly==5.24.1
plotly-resampler==0.10.0
pmdarima==2.0.4
preshed==3.0.9
prompt_toolkit==3.0.51
protobuf==5.28.3
psutil==6.1.0
pure_eval==0.2.3
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycaret==3.3.2
pydantic==2.9.2
pydantic_core==2.23.4
pydicom==3.0.1
Pygments==2.18.0
pyLDAvis==3.4.1
pynndescent==0.5.13
pynrrd==1.1.1
pyod==2.0.5
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.2
PyWavelets==1.8.0
pywin32==310
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
retrying==1.3.4
rich==13.9.3
rpds-py==0.24.0
rsa==4.9.1
scanorama==1.7.4
scanpy==1.11.1
schemdraw==0.15
scikit-base==0.7.8
scikit-image==0.24.0
scikit-learn==1.4.2
scikit-plot==0.3.7
scipy==1.11.4
seaborn==0.13.2
session-info2==0.1.2
setuptools==78.1.1
shellingham==1.5.4
six==1.17.0
sklearn-compat==0.1.3
sktime==0.26.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
spacy==3.8.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.36
sqlparse==0.5.3
srsly==2.5.1
stack-data==0.6.3
starlette==0.41.3
statsmodels==0.14.4
stringzilla==3.12.5
sympy==1.13.1
tangled-up-in-unicode==0.1.0
tbats==1.1.3
tenacity==9.1.2
termcolor==3.1.0
textblob==0.19.0
texttable==1.7.0
thinc==8.3.4
threadpoolctl==3.6.0
tifffile==2025.5.10
torch==2.5.1
torchvision==0.20.1
torchxrayvision==1.3.4
tornado==6.4.2
tqdm==4.67.0
traitlets==5.14.3
tsdownsample==0.1.4.1
typer==0.15.3
typing_extensions==4.13.2
tzdata==2025.2
umap==0.1.1
umap-learn==0.5.7
urllib3==2.4.0
uvicorn==0.32.1
visions==0.7.1
waitress==3.0.2
wasabi==1.1.3
watchfiles==1.0.5
wcwidth==0.2.13
weasel==0.4.1
websockets==15.0.1
Werkzeug==3.0.6
wheel==0.45.1
widgetsnbextension==4.0.14
wordcloud==1.9.4
wrapt==1.17.2
xgboost==3.0.0
xxhash==3.5.0
xyzservices==2025.4.0
yellowbrick==1.5
zipp==3.21.0
